safari-534.53-branch/LayoutTests/fast/files/create-blob-url-crash-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: webkitURL
Test that createObjectURL with no argument should not cause crash.
